+++ src/core/models/outputs/output_nft.h
@@ -66,9 +66,9 @@ size_t output_nft_serialize_len(output_nft_t* output);
  * @param[in] output NFT Output object.
  * @param[out] buf A buffer holds the serialized data
  * @param[in] buf_len The length of buffer
- * @return int 0 on success
+ * @return size_t The bytes written is returned, 0 on errors
  */
-int output_nft_serialize(output_nft_t* output, byte_t buf[], size_t buf_len);
+size_t output_nft_serialize(output_nft_t* output, byte_t buf[], size_t buf_len);
 
 /**
  * @brief Deserialize a binary data to a NFT Output object
